% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyMap.R
\name{CountyMap}
\alias{CountyMap}
\title{map county-level occurence of a species in any US state}
\usage{
CountyMap(
  dataframe,
  species_col,
  species_to_map,
  county_shapefile,
  state_col,
  state_to_map,
  crs_map
)
}
\arguments{
\item{dataframe}{an sf dataframe with a geometry column consisting of coordinates for each observation}

\item{species_col}{column in the dataframe identifying the species of each observation}

\item{species_to_map}{species to be mapped, wrap in quotes -> "Example sp."}

\item{county_shapefile}{a sf object of the counties of the United States}

\item{state_col}{state column in the counties object}

\item{state_to_map}{state that is to be mapped}

\item{crs_map}{CRS system to use for observation data and county shapefile}
}
\value{
a ggplot object
}
\description{
map county-level occurence of a species in any US state
}
\examples{
'CountyMap(frog_obs_iNat, Species, "Green Tree Frog", USAcounties, State, "GA", 4269)'
}
