% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyMap.R
\name{CountyMap_df}
\alias{CountyMap_df}
\title{produces sf datafrane with county level presence/absence data}
\usage{
CountyMap_df(
  dataframe,
  species_col,
  species_to_map,
  crs_map,
  county_shapefile,
  state_col,
  county_col,
  state_to_map,
  ...
)
}
\arguments{
\item{dataframe}{an sf dataframe with a geometry column consisting of coordinates for each observation}

\item{species_col}{column in the dataframe identifying the species of each observation}

\item{species_to_map}{species to be mapped, wrap in quotes -> "Example sp."}

\item{crs_map}{CRS to use for observation data and county shapefile}

\item{county_shapefile}{a sf object of the counties of the United States}

\item{state_col}{state column in the counties object (county_shapefile)}

\item{county_col}{county column in the counties object (county_shapefile)}

\item{state_to_map}{state that is to be mapped}

\item{...}{additional states to be mapped, optional}
}
\value{
a spatial dataframe
}
\description{
produces sf datafrane with county level presence/absence data
}
\examples{
'CountyMap_df(frog_obs_iNat, Species, "Green Tree Frog", 4269, USAcounties, State, "GA")'
}
