% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyMap.R
\name{CountyMap_map}
\alias{CountyMap_map}
\title{maps county level presence/absence data}
\usage{
CountyMap_map(
  dataframe,
  county_shapefile,
  crs_map,
  state_col,
  state_to_map,
  ...,
  color_present = "seagreen3",
  color_absent = "ivory1",
  color_background = "White",
  color_county_border = "grey41",
  color_state_border = "grey10"
)
}
\arguments{
\item{dataframe}{an sf dataframe with presence/absence data for each county, generated with CountyMap_df}

\item{county_shapefile}{a sf object of the counties of the United States}

\item{crs_map}{CRS to use for observation data and county shapefile, needs to match value inputted into CountyMap_df if two functions used in tandem}

\item{state_col}{state column in the counties object}

\item{state_to_map}{state that is to be mapped, wrap with ""}

\item{...}{additional states to be mapped, optional}

\item{color_present}{fill color for counties with species, defaults to green, wrap with ""}

\item{color_absent}{fill color for counties without species, defaults to ivory, wrap with ""}

\item{color_background}{fill color of background, defaults to white, wrap with ""}

\item{color_county_border}{color of county borders, defaults to light gray, wrap with ""}

\item{color_state_border}{color of state borders, defaults to dark gray, wrap with ""}
}
\value{
a ggplot object
}
\description{
maps county level presence/absence data
}
\examples{
'CountyMap_map(frog_pres, USAcounties, State, "GA", c("NC", "SC"))'
}
