% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_coords.R
\name{obs_to_coords}
\alias{obs_to_coords}
\title{convert a dataframe with lat and long columns to an sf object}
\usage{
obs_to_coords(df, coord_col, crs_add)
}
\arguments{
\item{df}{dataframe with coordinates column}

\item{coord_col}{column with coordinates}

\item{crs_add}{desired coordinate reference system}
}
\value{
simple features dataframe
}
\description{
convert a dataframe with lat and long columns to an sf object
}
\examples{
example.sf <- obs_to_coords(example.df, coords, 4269)
}
