% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeMap.R
\name{RangeMap}
\alias{RangeMap}
\title{map a polygon of species occurence}
\usage{
RangeMap(
  dataframe,
  species_col,
  species_to_map,
  background_map,
  greatlakes_map,
  crs_map
)
}
\arguments{
\item{dataframe}{an sf dataframe with a geometry column consisting of coordinates for each observation}

\item{species_col}{column in the dataframe identifying the species of each observation}

\item{species_to_map}{species to be mapped, wrap in quotes -> "Example sp."}

\item{background_map}{an sf object (map of a country, state, etc)}

\item{greatlakes_map}{an sf object of the Great Lakes}

\item{crs_map}{CRS system to use for observation data and background maps}
}
\value{
a ggplot object
}
\description{
map a polygon of species occurence
}
\examples{
'RangeMap(frog_obs_iNat, Species, "Green Tree Frog", NorthAmerica, GreatLakes, 26917)'
}
