% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeMap_2.R
\name{RangeMap_2}
\alias{RangeMap_2}
\title{Title}
\usage{
RangeMap_2(
  dataframe,
  species_col,
  species1,
  species2,
  background_map,
  greatlakes_map,
  crs_map
)
}
\arguments{
\item{dataframe}{an sf dataframe with a geometry column consisting of coordinates for each observation}

\item{species_col}{column in the dataframe identifying the species of each observation}

\item{species1}{1st species, wrap in quotes -> "Example sp."}

\item{species2}{2nd species, wrap in quotes -> "Example sp."}

\item{background_map}{an sf object (map of a country, state, etc)}

\item{greatlakes_map}{an sf object of the Great Lakes}

\item{crs_map}{CRS system to use for observation data and background maps}
}
\value{
ggplot object
}
\description{
Title
}
\examples{
`RangeMap_2(frog_obs_iNat, Species, "Green Tree Frog", "Gray Tree Frog", NorthAmerica, GreatLakes, 26917)`
}
