% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeMap.R
\name{RangeMap_map}
\alias{RangeMap_map}
\title{Produce range map for a species}
\usage{
RangeMap_map(
  dataframe,
  background_map,
  greatlakes_map,
  crs_map,
  color_background = "White",
  color_range = "forestgreen",
  alpha_range = 0.4
)
}
\arguments{
\item{dataframe}{an sf dataframe generated by TwoSpecies_df}

\item{background_map}{an sf object (map of a country, state, etc)}

\item{greatlakes_map}{an sf object of the Great Lakes}

\item{crs_map}{CRS to use for ranges and background maps}

\item{color_background}{fill color of background, defaults to white, wrap with ""}

\item{color_range}{fill color of range, defaults to green, wrap with ""}

\item{alpha_range}{transparency of range, defaults to 0.4}
}
\value{
ggplot object
}
\description{
Produce range map for a species
}
\examples{
'RangeMap_map(range, NorthAmerica, GreatLakes, 5070)'
}
