% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeMap_points.R
\name{RangeMap_points}
\alias{RangeMap_points}
\title{map the location of species observations}
\usage{
RangeMap_points(
  dataframe,
  species_col,
  species_to_map,
  background_map,
  greatlakes_map,
  crs_map,
  color_background = "White",
  color_points = "Black",
  alpha_points = 0.8
)
}
\arguments{
\item{dataframe}{an sf dataframe with a geometry column consisting of coordinates for each observation}

\item{species_col}{column in the dataframe identifying the species of each observation}

\item{species_to_map}{species to be mapped, wrap in quotes -> "Example sp."}

\item{background_map}{an sf object (map of a country, state, etc)}

\item{greatlakes_map}{an sf object of the Great Lakes}

\item{crs_map}{CRS system to use for observation data and background maps}

\item{color_background}{fill color of background, defaults to white, wrap with ""}

\item{color_points}{color of points, defaults to black, wrap with ""}

\item{alpha_points}{transparency of points, defaults to 0.8}
}
\value{
a ggplot object
}
\description{
map the location of species observations
}
\examples{
'RangeMap_points(frog_obs_iNat, Species, "Green Tree Frog", NorthAmerica, GreatLakes, 26917)'
}
