% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpecies_Map.R
\name{TwoSpecies_df}
\alias{TwoSpecies_df}
\title{Produces sf dataframe with range polygons for two species}
\usage{
TwoSpecies_df(
  dataframe,
  species_col,
  species1,
  species2,
  background_map,
  greatlakes_map,
  crs_map
)
}
\arguments{
\item{dataframe}{an sf dataframe with a geometry column consisting of coordinates for each observation}

\item{species_col}{column in the dataframe identifying the species of each observation}

\item{species1}{1st species, wrap in quotes -> "Example sp."}

\item{species2}{2nd species, wrap in quotes -> "Example sp."}

\item{background_map}{an sf object (map of a country, state, etc)}

\item{greatlakes_map}{an sf object of the Great Lakes}

\item{crs_map}{CRS to use for observation data and background maps}
}
\value{
a spatial dataframe
}
\description{
Produces sf dataframe with range polygons for two species
}
\examples{
'TwoSpecies_df(frog_obs, Species, "Green", "Frog", NorthAmerica, GreatLakes, 26917)'
}
