% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_sf.R
\name{view_sf}
\alias{view_sf}
\title{view a spatial dataframe in the R studio viewer}
\usage{
view_sf(df)
}
\arguments{
\item{df}{an sf dataframe with a geometry column}
}
\value{
nothing
}
\description{
view a spatial dataframe in the R studio viewer
}
\examples{
'view_sf(USA_counties)'
}
